Working on Delete and Update Jonral enrities @DeleteMapping(id/{myId})

+++ src/main/java/net/engineeringdigest/journalApp/Service/JournalEntryService.java
@@ -4,6 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import net.engineeringdigest.journalApp.Entity.JournalEntry;
 import net.engineeringdigest.journalApp.Entity.User;
 import net.engineeringdigest.journalApp.Repository.JournalEntryRepository;
+import net.engineeringdigest.journalApp.Repository.UserRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -17,6 +18,8 @@ import java.util.Optional;
 @Slf4j
 @Component
 public class JournalEntryService {
+    @Autowired
+    private UserRepository userRepository;
 
     @Autowired
     private JournalEntryRepository journalEntryRepository;
@@ -25,21 +28,21 @@ public class JournalEntryService {
 
     /*Create/Save Entry*/
     @Transactional
-    public void saveEntry(JournalEntry journalEntry, String userName) {
+    public void saveNewUser(JournalEntry journalEntry, String userName) {
         try {
             User user = userService.findByUserName(userName);
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e) {
             log.error("Exception", e);
-            throw  new RuntimeException("An error occurred while saving the journal entry", e);
+            throw new RuntimeException("An error occurred while saving the journal entry", e);
         }
     }
 
     /*This Method Used For Update The Current Data */
-    public void saveEntry(JournalEntry journalEntry) {
+    public void saveNewUser(JournalEntry journalEntry) {
         try {
             journalEntryRepository.save(journalEntry);
         } catch (Exception e) {
@@ -56,13 +59,26 @@ public class JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId id) {
         return journalEntryRepository.findById(id);
     }
-    /*Find by ID's*/
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(entry -> entry.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    /*Delete by ID's*/
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(entry -> entry.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the journal entry" ,e);
+        }
+        return removed;
+    }
+
+    public User findByUserName(String userName) {
+        return userRepository.findByUserName(userName);
     }
 }
-// Web Security Time 9:13 TODO Remember
